chore: move view files to a browser folder
